+++ src/games/takesix/config.ts
@@ -5,7 +5,6 @@ import { Board } from './board';
 const config: IGameConfig = {
   bgioGame: TakeSixGame,
   bgioBoard: Board,
-  debug: false,
 };
 
 export default config;
